Authorization was done!

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,7 @@ use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller implements HasMiddleware
@@ -42,7 +43,7 @@ class PostController extends Controller implements HasMiddleware
             $path = $request->file('photo')->storeAs('post-photos', $name);
         }
         $post = Post::create([
-            'user_id' => auth()->user()->id(),
+            'user_id' => auth()->id(),
             'category_id' => $request->get('category_id'),
             'title' => $request -> get('title'),
             'short_content' => $request -> get('short_content'),
@@ -69,11 +70,15 @@ class PostController extends Controller implements HasMiddleware
 
     public function edit(Post $post)
     {
+        Gate::authorize('update', $post);
+
         return view('posts.edit')->with(['post' => $post]);
     }
 
     public function update(StorePostRequest $request, Post $post)
     {
+        Gate::authorize('update', $post);
+
         if ($request->hasFile('photo')) {
             if (isset($post->photo)) {
                 Storage::delete($post->photo);
